Allow to remove molecules in a sphere
Remove all water molecules (or any residue given as water molecule) if
at least one atom of the water molecule is at less than a given distance
to the geometrical center of group of atoms.

The reference group of atoms is defined by giving a list of residue
names of interest with the --sphere option. The radius of the sphere is
hiven using --radius.